+++ test/middleware.test.js
@@ -1,7 +1,9 @@
-var bogart = require('../lib/bogart')
-  , Q      = require("q")
-  , assert = require('assert')
-  , path   = require('path');
+var bogart    = require('../lib/bogart')
+  , Q         = require("q")
+  , assert    = require('assert')
+  , path      = require('path')
+  , fs        = require('fs')
+  , security  = require("../lib/security");
 
 exports["test parses JSON"] = function(beforeExit) {
   var forEachDeferred = Q.defer()
@@ -122,4 +124,187 @@ exports["test gzip downloads as text/html"] = function(beforeExit) {
     assert.equal(200, response.status);
     assert.equal('text/html', response.headers['content-type']);
   });
+};
+
+exports["test error middleware has default response when error is thrown"] = function(beforeExit) {
+  var response = null
+    , app      = new bogart.middleware.Error(function(req) { throw new Error('intentional'); });
+  
+  Q.when(app({ method: 'GET', env: {}, headers: {}, pathInfo: '/' }), function(resp) {
+    response = resp;
+  });
+
+  beforeExit(function() {
+    assert.isNotNull(response);
+    assert.equal(500, response.status);
+    assert.equal('text/html', response.headers['content-type']);
+  });
+};
+
+exports["test flash"] = function(beforeExit) {
+  var app
+    , headers = { 'content-type': 'text/plain' }
+    , request = { headers: headers, body:[] }
+    , foo
+    , cookieStr;
+    
+  app = bogart.middleware.Flash({}, function(req) {
+    req.flash("foo", "bar");
+
+    foo = req.flash("foo");
+    return {
+      status: 200,
+      body: [],
+    }
+  });
+
+  var initialResp = app(request);
+  Q.when(app(request), function(resp) {
+    cookieStr = resp.headers["Set-Cookie"].join("").replace(/;$/, "");
+
+    assert.equal(foo, 'bar');
+  });
+  
+  request.headers.cookie = cookieStr;
+  var secondResp = app(request);
+
+  beforeExit(function() {
+    // the first attempt to retrieve "foo" should be undefined
+    assert.isUndefined(foo);
+  });
+};
+
+
+exports["test error middleware has default response when promise is rejected"] = function(beforeExit) {
+  var response = null
+    , app      = new bogart.middleware.Error(function(req) { return require('q').reject('rejected'); });
+  
+  Q.when(app({ method: 'GET', env: {}, headers: {}, pathInfo: '/' }), function(resp) {
+    response = resp;
+  });
+
+  beforeExit(function() {
+    assert.isNotNull(response);
+    assert.equal(500, response.status);
+    assert.equal('text/html', response.headers['content-type']);
+  });
+};
+
+exports["test parted json"] = function(beforeExit) {
+  var request       = null
+    , parted        = new bogart.middleware.Parted(function(req) { request = req; return {}; });
+  
+  response = parted({
+    method: 'POST',
+    env: {},
+    headers: { 'content-type': 'application/json' },
+    body: [ '{ "hello": "world" }' ]
+  });
+
+  beforeExit(function() {
+    assert.isNotNull(request);
+    assert.isNotNull(request.body);
+    assert.equal('object', typeof request.body);
+    assert.equal('world', request.body.hello);
+  });
+};
+
+exports["test parted form"] = function(beforeExit) {
+  var request     = null
+    , parted      = new bogart.middleware.Parted(function(req) { request = req; return {}; })
+    , body        = {}
+    , bodyDefer   = require('q').defer();
+  
+  body.forEach = function(callback) {
+    callback('hello=one&hello=two');
+
+    return bodyDefer.promise;
+  };
+  
+  response = parted({
+    method: 'POST',
+    env: {},
+    headers: { 'content-type': 'application/x-www-form-urlencoded' },
+    body: body
+  });
+
+  bodyDefer.resolve();
+
+  beforeExit(function() {
+    assert.isNotNull(request);
+    assert.ok(!!request.body);
+    assert.ok(!!request.body.hello);
+    assert.equal(2, request.body.hello.length);
+  });
+};
+
+exports["test parted multipart"] = function(beforeExit) {
+  var request = null
+    , parted  = new bogart.middleware.Parted(function(req) { request = req; return {}; });
+  
+  fs.readFileSync(path.join(__dirname, 'fixtures', 'chrome.part'));
+  
+  response = parted(multipartRequest(100, 'chrome'));
+
+  beforeExit(function() {
+    assert.ok(!!request.body);
+    assert.ok(!!request.body.content, 'No file path');
+  });
+};
+
+/**
+ * Create a mock request
+ * 
+ * Modified from the mock request method in Parted in compliance with the license.
+ */
+function multipartRequest(size, file) {
+  file = path.join(__dirname, 'fixtures', file + '.part');
+
+  var stream = fs.createReadStream(file, {
+    bufferSize: size
+  });
+
+  var boundary = fs
+    .readFileSync(file)
+    .toString('utf8')
+    .match(/--[^\r\n]+/)[0]
+    .slice(2);
+
+  return {
+    headers: {
+      'content-type': 'multipart/form-data; boundary="' + boundary + '"'
+    },
+    method: 'POST',
+    env: {},
+    pipe: function(dest) {
+      stream.pipe(dest);
+    },
+    emit: function(ev, err) {
+      if (ev === 'error') this.errback && this.errback(err);
+      return this;
+    },
+    on: function(ev, func) {
+      if (ev === 'error') this.errback = func;
+      return this;
+    },
+    destroy: function() {
+      stream.destroy();
+      return this;
+    },
+    body: {
+      forEach: function(fn) {
+        var deferred = Q.defer();
+
+        stream.on('data', function(data) {
+          fn(data);
+        });
+
+        stream.on('end', function() {
+          deferred.resolve();
+        });
+
+        return deferred.promise;
+      }
+    }
+  };
 };